backend/requirements.txt
Flask==2.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.7.0
gunicorn==20.1.0
python-dateutil==2.8.2
